requirements.txt
fastapi
uvicorn
openai
pinecone-client
python-multipart
ffmpeg-python
pydub
tqdm
pillow
pytesseract
whisper
python-dotenv
